+++ src/ui/ui_base.h
@@ -19,7 +19,7 @@ public:
 
 void display_grid(GxEPD2_GFX& display);
 
-void display_text_center(GxEPD2_GFX& display, int16_t x, int16_t y, const char *string, int16_t w = 0, int16_t h = 0, const GFXfont *font = &PAPERDINK_FONT_SML, uint8_t color = GxEPD_BLACK, uint8_t size = 1);
+void display_text_center(GxEPD2_GFX& display, int16_t x, int16_t y, const char *string, int16_t w = 0, int16_t h = 0, const GFXfont *font = &PAPERDINK_FONT_SML, uint16_t color = GxEPD_BLACK, uint8_t size = 1);
 
 
 uint16_t read16(File& f);
